Add ownership to 'tip' commands.
Currently, only Radar can !forget commands. This leads to a lot of
outdated and useless information in the 'tip' commands for railsbot. It
would be nice if the person who added the command also had the ability
to forget and re-add commands in case information gets updated over
time. This change (I think) enables that, but some kind of DB schema
update will have to be run on the railsbot server in order to actually
make it happen.